--- jyx2/Assets/Scripts/BattleManager/BattleConfig.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f83ce2e801885b4280f3328af2af751
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
